requirements.txt
streamlit
python-dotenv
google-generativeai
streamlit-chat
audio_recorder_streamlit
pydub
speechrecognition